[CHORE] Update release workflow status to 'draft'
This commit modifies the GitHub Actions release workflow to set the status of new Play Store releases to `draft` instead of using `changesNotSentForReview: true`.

This change updates the `upload-google-play` action's configuration, ensuring that new releases are created as drafts in the Google Play Console, allowing for manual review and rollout.

Signed-off-by: Zakir Sheikh <sheikhzakirahmad@gmail.com>